requirements.txt
# Core dependencies
PyTDC>=0.4.0
rdkit>=2023.9.5
tqdm>=4.66.1

# Machine learning
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Scientific computing
scipy>=1.10.0